modify commandrole

+++ sources/cmds/commandrole.py
@@ -8,7 +8,7 @@ from datahook import yamlhook
 
 ydata = yamlhook("config.yaml")
 
-class commandsrole(commands.Cog):
+class commandrole(commands.Cog):
 
     def __init__(self,bot:commands.Bot):
         self.bot = bot
@@ -17,7 +17,7 @@ class commandsrole(commands.Cog):
 
         if("commandrole" not in dict(self.rolesdata).keys()):
             self.dataOperate(dictTopic="commandrole", setting={})
-            # reload rolesdata
+            # upload rolesdata
             self.rolesdata = ydata.load()
 
     def dataReload(self):
@@ -26,7 +26,8 @@ class commandsrole(commands.Cog):
     @commands.Cog.listener()
     async def on_message(self,msg:discord.Message):
         if(msg.channel.id in dict(self.rolesdata)["commandrole"].keys()):
-            await msg.delete()
+            if(msg.author.bot == False):
+                await msg.delete()
             _channel = self.rolesdata["commandrole"][msg.channel.id]
             if(msg.content in _channel["command"]):
                 try:  
@@ -41,7 +42,7 @@ class commandsrole(commands.Cog):
     @commands.command(name='addcm', help="新增指令身分組 `指令` `身分組`")
     @commands.has_permissions(administrator=True,manage_roles=True)
     async def add_command_role(self, ctx:commands.Context, cmds, * , role:discord.Role):
-
+        # split keyword
         split_cmd = str(cmds).split(',')
         if(ctx.channel.id not in dict(self.rolesdata["commandrole"]).keys()):
             newData = {
@@ -51,19 +52,34 @@ class commandsrole(commands.Cog):
                 }
             }
             self.dataOperate(dictTopic="commandrole", setting=newData)
-            # update role data
-            self.rolesdata = ydata.load()
             # create successful embed message
             successEmbed = discord.Embed(title="成功!",description="已成功在此頻道新增關鍵字偵測",colour=self.bot.default_colour)
             successEmbed.add_field(name="使用指令", value=str(cmds).replace(',','/'), inline=False)
             successEmbed.add_field(name="獲取身分", value=role.mention,inline=False)
             msgOut = await ctx.send(embed=successEmbed)
-
+            # wait 5 seconds
             await asyncio.sleep(5)
+            # delete message
+            await msgOut.delete()
+            # send remind message
+            await ctx.send("```md\n- 請觀看規則之後輸入相關指令獲取身分組喔!\n```")
+            # update role data
+            self.rolesdata = ydata.load()
+        else:
+            existEmbed = discord.Embed(title="提醒!",description="每個頻道只能有一個指令而已喔。",colour=self.bot.default_colour)
 
+            # _channel = self.rolesdata["commandrole"][ctx.channel.id]
+            # target_role = get(ctx.author.guild.roles, id=_channel["role"])
+            # use_command = ""
+            # existEmbed.add_field(name="使用指令", value=use_command.join(_channel["command"]), inline=False)
+            # existEmbed.add_field(name="獲取身分", value=target_role.mention,inline=False)
+
+            # send embed message
+            msgOut = await ctx.send(embed=existEmbed)
+            # wait 5 seconds
+            await asyncio.sleep(5)
+            # delete message
             await msgOut.delete()
 
-            await ctx.send("```md\n- 請觀看規則之後輸入相關指令獲取身分組喔!\n```")
-
 def setup(bot):
-    bot.add_cog(commandsrole(bot))
+    bot.add_cog(commandrole(bot))
